Refactor images and files IV
